Docker images v2 (#1021)
New naming scheme appends `-v#` to the image name. The tag part is not changed
(so you can still tell the difference between Ubuntu versions)

Removed Docker Hub pushes. Now we use `ghcr.io` for both images.

We no longer set `CC` and `CXX` in the image. We now use the default compiler
(GCC_ for cross-compilation on because Ubuntu and vcpkg have better support for
cross-compilation when using GCC.